refactoring and documentation
