+++ src/main/kotlin/io/github/graphglue/definition/ManyRelationshipDefinition.kt
@@ -2,7 +2,7 @@ package io.github.graphglue.definition
 
 import graphql.schema.GraphQLFieldDefinition
 import io.github.graphglue.connection.generateConnectionFieldDefinition
-import io.github.graphglue.graphql.SchemaTransformationContext
+import io.github.graphglue.graphql.schema.SchemaTransformationContext
 import io.github.graphglue.model.Direction
 import io.github.graphglue.model.Node
 import kotlin.reflect.KClass
