+++ src/main/kotlin/io/github/graphglue/definition/RelationshipDefinition.kt
@@ -2,12 +2,10 @@ package io.github.graphglue.definition
 
 import com.expediagroup.graphql.generator.annotations.GraphQLDescription
 import com.expediagroup.graphql.generator.annotations.GraphQLIgnore
-import graphql.language.DirectiveDefinition
-import graphql.schema.GraphQLDirective
 import graphql.schema.GraphQLFieldDefinition
 import io.github.graphglue.data.execution.NodeQueryResult
 import io.github.graphglue.data.repositories.RelationshipDiff
-import io.github.graphglue.graphql.SchemaTransformationContext
+import io.github.graphglue.graphql.schema.SchemaTransformationContext
 import io.github.graphglue.graphql.extensions.getPropertyName
 import io.github.graphglue.model.Direction
 import io.github.graphglue.model.Node
